Move Transaction schema to ledger module
