+++ src/ledger/books.ts
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import { Transaction } from "./schema"
 // we have to import schema before medici
 import { book } from "medici"
